complete bignum division :^)
This is a big milestone - I can finally do the DH math in a reasonable
amount of time (and RSA too, but I haven't tested that yet)! In other
words, the bignum library is finally complete enough to support our SSH
implementation!

btw, this took over a month of work '^^